move types to lib/skills instead of lib/api
